# cli/source/lib/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: string;
};

export type LlmConfigDto = {
    /**
     * LLM provider type
     */
    type?: 'cloud' | 'self-hosted';
    /**
     * Cloud provider name (anthropic, openai, groq)
     */
    provider?: string;
    /**
     * Model name
     */
    model?: string;
    /**
     * Base URL for self-hosted LLM
     */
    baseURL?: string;
    /**
     * API key for authentication
     */
    apiKey?: string;
};

export type ConfigDto = {
    /**
     * LLM configuration
     */
    llm?: LlmConfigDto;
};

export type AppControllerGetHelloData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/';
};

export type AppControllerGetHelloResponses = {
    200: unknown;
};

export type ConfigControllerResetConfigData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/config';
};

export type ConfigControllerResetConfigResponses = {
    default: ConfigDto;
};

export type ConfigControllerResetConfigResponse = ConfigControllerResetConfigResponses[keyof ConfigControllerResetConfigResponses];

export type ConfigControllerGetConfigData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/config';
};

export type ConfigControllerGetConfigResponses = {
    default: ConfigDto;
};

export type ConfigControllerGetConfigResponse = ConfigControllerGetConfigResponses[keyof ConfigControllerGetConfigResponses];

export type ConfigControllerPartialUpdateData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/config';
};

export type ConfigControllerPartialUpdateResponses = {
    default: ConfigDto;
};

export type ConfigControllerPartialUpdateResponse = ConfigControllerPartialUpdateResponses[keyof ConfigControllerPartialUpdateResponses];

export type ConfigControllerUpdateConfigData = {
    body: ConfigDto;
    path?: never;
    query?: never;
    url: '/config';
};

export type ConfigControllerUpdateConfigResponses = {
    default: ConfigDto;
};

export type ConfigControllerUpdateConfigResponse = ConfigControllerUpdateConfigResponses[keyof ConfigControllerUpdateConfigResponses];